switch to meson to build mpv
This provides us with a mpv(1) binary which the previous build
didn't. This is going to be useful to debug playback issues, but it's
also simpler as we don't need to download WAF to get the build to
work, so one less dependency to track.

If we do want to remove the mpv binary later, we can add it to the
cleanup target.